+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ```
 IMP: Screen: Use fewer css classes
+IMP: Support last version of Odoo (master branch)
+
+ADD: Command 'barcode': Simulate barcode scanning and info
 
 FIX: Command 'lang': UtilsBackend is undefined (issue #55)
 ```